% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{This function says hello to people (Header)
subheader
text}
\usage{
hello(name = "Julia")
}
\arguments{
\item{name}{A string of parameters preferentiallz name of a person. (Default = "Julia")}
}
\value{
prints Hello name to screen
}
\description{
This function says hello to people (Header)
subheader
text
}
\examples{
hello("Julia")
}
